Fix deletion bugs + add error checks on catalog deletion in service

+++ server/src/catalog/catalog.service.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Inject, Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
+import { In, Repository } from "typeorm";
 import { Catalog } from "./models/catalog.entity";
 import { ArchiveService } from "../archive/archive.service";
 import { CatalogFilterDto } from "./models/catalog-filter.dto";
@@ -30,21 +30,53 @@ export class CatalogService {
         return this.catalogRepository.update(id, data);
     }
 
-    async delete(id: number): Promise<any> {
-        const backup = await this.findOne(id);
+    /*async deleteFromObjectId(object_id: string): Promise<any> {
+        return this.catalogRepository.delete({ object_id: object_id });
+    }*/
+
+    async delete(id: number): Promise<Catalog> {
+        const backup = await this.findOne({ id: id });
+
+        // Create archive
+        const archive = await this.archiveService.create(backup);
+
+        // Check if archive has successfully been created
+        if (!archive) {
+            return null;
+        }
+
+        // Check if object has successfully been backed up correctly
+        if (backup.object_id !== archive.object_id) {
+            return null;
+        }
+
         await this.catalogRepository.delete(id);
-        await this.archiveService.create(backup);
+
         return backup;
     }
 
-    async deleteFromObjectId(object_id: string): Promise<any> {
-        return this.catalogRepository.delete({ object_id: object_id });
-    }
-
     async deleteArray(ids: number[]) {
-        const backup = await this.catalogRepository.findByIds(ids);
-        await this.catalogRepository.delete(ids);
-        await this.archiveService.create(backup);
+        const backup = await this.catalogRepository.findBy({ id: In(ids) });
+
+        // Create archive for each backup item
+        for (const item of backup) {
+            // Create archive
+            const archive = await this.archiveService.create(item);
+
+            // Check if archive has successfully been created
+            if (!archive) {
+                return null;
+            }
+
+            // Check if object has successfully been backed up correctly
+            if (item.object_id !== archive.object_id) {
+                return null;
+            }
+
+            // Delete from catalog if archive was created
+            await this.catalogRepository.delete(item.id);
+        }
+
         return backup;
     }
 
@@ -52,10 +84,34 @@ export class CatalogService {
         const backup = await this.catalogRepository.find({
             where: { company: company_id }
         });
-        await this.catalogRepository.delete({ company: company_id });
-        await this.archiveService.create(backup);
+
+        // Create archive for each backup item
+        for (const item of backup) {
+            // Create archive
+            const archive = await this.archiveService.create(item);
+
+            // Check if archive has successfully been created
+            if (!archive) {
+                return null;
+            }
+
+            // Check if object has successfully been backed up correctly
+            if (item.object_id !== archive.object_id) {
+                return null;
+            }
+
+            // Delete from catalog if archive was created
+            await this.catalogRepository.delete(item.id);
+        }
+
+        // Make sure everything has been deleted (not necessary)
+        await this.catalogRepository.delete({
+            company: company_id
+        });
+
         return backup;
     }
+
     async filterCatalog(filterDto: CatalogFilterDto): Promise<Catalog[]> {
         // Implement your filtering logic based on the provided criteria
         return this.catalogRepository.find({
